requirements.txt
torch==2.0.1+cu117
numpy==1.21.5
easydict==1.10
transformers==4.29.2
peft==0.3.0
accelerate==0.19.0
PyYAML==6.0
pandas==1.4.2
datasets==2.12.0
evaluate==0.4.0
scikit-learn==1.0.2
tqdm==4.64.0